Начал реализовывать экран отображения тела работы. Сделал выбор клиента....

+++ app/src/main/java/com/example/hairmasterplaner/domain/jobBody/JobBodyItem.kt
@@ -2,7 +2,7 @@ package com.example.hairmasterplaner.domain.jobBody
 
 data class JobBodyItem(
     val id:Long,
-    val jobId:Int,
+    val jobId:Long,
     val jobElementItemId: Int,
     val amount:Int?,
     val price:Int,
